feat: move employment output to data_access dir
to be able to add authorization policy on employment authorized clients
